integrate(6750): verify http transport test extraction entrypoint
