+++ crates/kamn-core/tests/kolme_runtime_commit_http_transport/support/http_server_support.rs
@@ -4,6 +4,9 @@ use super::*;
 mod request_read_support;
 #[path = "http_server_support/response_server_support.rs"]
 mod response_server_support;
+#[path = "http_server_support/keep_alive_support.rs"]
+mod keep_alive_support;
 
 pub(crate) use request_read_support::*;
 pub(crate) use response_server_support::*;
+pub(crate) use keep_alive_support::*;
